# Speed up terminal reconnection