… extensions/vscode-api-tests/src/singlefolder-tests/terminal.test.ts
… src/vs/workbench/contrib/terminal/electron-sandbox/localPty.ts
… src/vs/workbench/contrib/terminal/electron-sandbox/localTerminalBackend.ts