- [ ] Merge https://github.com/microsoft/vscode/pull/186041 which this builds upon

Summary:

- Terminals are all re-created synchronously, this speeds up reconnection a little, but the biggest win here is the user should see all the restored terminals appear all at once instead of trickling in
- Added replay terminal stats
- Remove orphan check when attaching to processes (> 50% speed of replay)
  - Before (10 terminals): `willReplay -> didReplay`
    - Run 1: 7504ms
    - Run 2: 7495ms
  - After (10 terminals): `willReplay -> didReplay`
    - Run 1: 3433ms
    - Run 2: 3545ms
- The removal of the orphan check _may_ fix #133542, I want to verify more before closing that though

Fixes https://github.com/microsoft/vscode/issues/186068
Fixes https://github.com/microsoft/vscode/issues/186069
Part of https://github.com/microsoft/vscode/issues/185393
Part of #133542